Small fixes to various property types

+++ src/components/Property/Types/ListProperty/IntNumberListProperty.tsx
@@ -1,5 +1,5 @@
 import { ListProperty, ListPropertyProps } from './ListProperty';
 
 export function FloatingNumberListProperty(props: ListPropertyProps) {
-  return <ListProperty {...props} valueType={'float'} />;
+  return <ListProperty {...props} valueType={'int'} />;
 }
